requirements.txt
requests[socks]
beautifulsoup4
pyyaml
flask
gunicorn
bech32
pycardano
blockfrost-python
groq